[OOT] Support sync_model_loading for OOT (#25126)
